Make hashing faster
After we benchmark the code we found out that the bottleneck is still
hashing, even after we changed the hashing algorithm to ahash. We are
making ahash job easier by transforming the whole state in a single
integer, feeding it to ahash.

This give us 20% performance boost in min-max algorithm.